pyproject.toml
[project]
name = "qt-ingot"
version = "0.0.1"
description = "A lightweight, themeable boilerplate for creating tab-based PyQt applications."
readme = "README.md"
keywords = ["qt", "pyqt", "pyqt6", "gui", "template", "scaffolding"]
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Fernando Bryan Reza Campos", email = "fer.rezac@outlook.com"},
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Desktop Environment",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "PyQt6>=6.9.1",
    "libsass>=0.23.0",
    "rune-lib>=0.0.2",
]

[project.urls]
Homepage = "https://github.com/Yrrrrrf/qt-ingot"
Repository = "https://github.com/Yrrrrrf/qt-ingot"


[tool.hatch.build.targets.wheel]
packages = ["src/ingot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
